+++ src/app.rs
@@ -4,19 +4,22 @@ use std::sync::Arc;
 use anyhow::Result;
 use threadpool::ThreadPool;
 
-use crate::pipeline::{Message, Server};
+use crate::params::Params;
+use crate::server::{Message, Server};
 use crate::tui::Tui;
 
 pub struct App {
     tpool: ThreadPool,
     server: Arc<Server>,
+    app_opts: Arc<Params>,
 }
 
 impl App {
-    pub fn new() -> Self {
+    pub fn new(app_opts: Arc<Params>) -> Self {
         Self {
             tpool: ThreadPool::new(8),
             server: Arc::new(Server::new().expect("server init failed")),
+            app_opts,
         }
     }
 
@@ -25,7 +28,8 @@ impl App {
         let (app_tx, app_rx) = channel::<Message>();
 
         let server_ptr = self.server.clone();
-        let ui = Tui::new(app_tx);
+        let tui_server_ptr = self.server.clone();
+        let ui = Tui::new(app_tx, tui_server_ptr);
 
         self.tpool.execute(move || {
             server_ptr.start(server_rx).expect("server init failed");
