+++ src/main.rs
@@ -8,20 +8,21 @@ mod scanner;
 
 /* version 2.0 modules*/
 mod cli;
-mod pipeline;
+mod server;
 mod tui;
 
 use anyhow::Result;
 
 use self::app::App;
-// use clap::Parser;
+use clap::Parser;
+use params::Params;
+use std::sync::Arc;
 // use formatter::Formatter;
-// use params::Params;
 // use processor::Processor;
 // use scanner::Scanner;
 
 fn main() -> Result<()> {
-    // let app_args = Params::parse();
+    let app_args = Params::parse();
     // let scan_results = Scanner::build(&app_args)?.scan()?;
     // let processor = Processor::new(scan_results);
     // let results = processor.sizewise()?.hashwise()?;
@@ -31,7 +32,9 @@ fn main() -> Result<()> {
     //     true => interactive::init(results.files, &app_args)?,
     // }
 
-    App::new().start().expect("app init failed.");
+    App::new(Arc::new(app_args))
+        .start()
+        .expect("app init failed.");
 
     Ok(())
 }
