+++ src/tui/mod.rs
@@ -6,15 +6,17 @@ use ratatui::crossterm::event::{self, Event, KeyCode};
 use ratatui::widgets::Paragraph;
 use ratatui::{DefaultTerminal, Frame};
 
-use crate::pipeline::Message;
+use crate::server::{Message, Server};
+use std::sync::Arc;
 
 pub struct Tui {
     app_tx: Sender<Message>,
+    server: Arc<Server>,
 }
 
 impl Tui {
-    pub fn new(app_tx: Sender<Message>) -> Self {
-        Self { app_tx }
+    pub fn new(app_tx: Sender<Message>, server: Arc<Server>) -> Self {
+        Self { app_tx, server }
     }
 
     pub fn start(&self) -> Result<()> {
@@ -26,7 +28,7 @@ impl Tui {
     }
 
     fn poll_events() -> Result<Message> {
-        match event::poll(Duration::from_millis(250)).context("event polling failed.")? {
+        match event::poll(Duration::from_millis(50)).context("event polling failed.")? {
             true => match event::read().context("event read failed.")? {
                 Event::Key(key) => match key.code {
                     KeyCode::Char('q') => Ok(Message::Exit),
@@ -39,8 +41,6 @@ impl Tui {
     }
 
     fn handle_events(&self) -> Result<Message> {
-        let handleable_event = Self::poll_events();
-
         match Self::poll_events() {
             Ok(Message::Exit) => {
                 self.app_tx
@@ -52,13 +52,13 @@ impl Tui {
         }
     }
 
-    fn draw(frame: &mut Frame) {
+    fn draw(&self, frame: &mut Frame) {
         frame.render_widget(Paragraph::new("Hello, World"), frame.area());
     }
 
     fn run(&self, mut terminal: DefaultTerminal) -> Result<()> {
         loop {
-            terminal.draw(Self::draw)?;
+            terminal.draw(|f| self.draw(f))?;
             match self.handle_events() {
                 Ok(Message::Exit) => break,
                 _ => {}
